Se ajusta edicion.

+++ Usuarios.Entidades/Persona.cs
@@ -2,14 +2,14 @@
 {
     public class Persona
     {
-        public int IdUsuario { get; set; }
+       public int IdUsuario { get; set; }
         public string Nombres { get; set; }
         public string Apellidos { get; set; }
         public string Edad { get; set; }
         public string NumeroDocumento { get; set; }
         public int IdTipoDocumento { get; set; }        
         public int IdCiudadResidencia { get; set; }    
-        public int IdDepartamentoResidencia { get; set; }       
+        public decimal IdDepartamentoResidencia { get; set; }       
         public string Direccion { get; set; }
         public string Telefono { get; set; }
         public string Celular { get; set; }
